pyproject.toml
[tool.poetry]
name = "picpickstudio-imageautoloader"
version = "0.1.0"
description = ""
authors = ["Yu Li <ylilarry@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10"
Pillow = "^9.0.1"
watchdog = "^2.1.6"
PyYAML = "^6.0"
click = "^8.0.4"
marshmallow-dataclass = "^8.5.3"
marshmallow = "^3.15.0"
pywin32 = {version = "^303", platform = "windows"}

[tool.poetry.dev-dependencies]
black = "^22.1.0"
mypy = "^0.941"
types-PyYAML = "^6.0.5"
ipython = "^8.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
